+++ Return.h
@@ -15,7 +15,7 @@ public:
   ~Return();
 
   void add(string dr, double amount, string returner, string bc, string dB);
-  void updateReturn(double amount, string newDateReturned);
+  void updateReturn(int amount, string newDateReturned);
   int getReturn();
   string getDatePurchased();
   string getCompany();
