+++ models/History.cpp
@@ -187,7 +187,8 @@ string History::getCompany(){
     return company;
 }
 
-int History::getDifference(){
+// The change in units
+int History::getDifference() {
   return difference;
 }
 
@@ -195,31 +196,31 @@ int History::getDumped(){
   return dumped;
 }
 
-int History::getReturned(){
+int History::getReturned() {
   return returned;
 }
 
-int History::getTui(){
+int History::getTui() {
     return mTui;
 }
 
-string History::getPrice(){
+string History::getPrice() {
     return price;
 }
 
-string History::getDatePurchased(){
+string History::getDatePurchased() {
     return dayPurchased;
 }
 
-string History::getDateSold(){
+string History::getDateSold() {
     return daySold;
 }
 
-string History::getDateToCompare(){
+string History::getDateToCompare() {
     return dateToCompare;
 }
 
-void History::buy(double amount, string bc, string date, string p){
+void History::buy(int amount, string bc, string date, string p){
 
   difference = amount;
   company = bc;
@@ -229,7 +230,7 @@ void History::buy(double amount, string bc, string date, string p){
   type = "Buy";
 }
 
-void History::sell(double amount, string b, string date, string p, string dp,string c , int change){
+void History::sell(int amount, string b, string date, string p, string dp,string c , int change){
 
   difference = amount *(-1);
   customer = b;
@@ -242,7 +243,7 @@ void History::sell(double amount, string b, string date, string p, string dp,str
   type = "Sell";
 }
 
-void History::dump(double amount, string date, string bd, string comp, int change){
+void History::dump(int amount, string date, string bd, string comp, int change){
     difference = amount *(-1);
     daySold = date;
     dumped = 1;
@@ -269,7 +270,7 @@ void History::returnn(string dR, double amount, string returner, string c, strin
 
 // Return to farm
 // - paramenter date: The day that the "tui" happened
-void History::tui(double amount, string dateTui, string datePurchased, string comp, int change) {
+void History::tui(int amount, string dateTui, string datePurchased, string comp, int change) {
     difference = amount *(-1);
     daySold = dateTui;
     mTui = 1;
